refactor(frontend): rename agent studio composer internals
