+++ packages/frontend/src/pages/agents/composer-runtime/use-context-usage.ts
@@ -8,7 +8,7 @@ import {
   extractLatestContextUsage,
   extractLatestContextUsageEntry,
   toModelDescriptorByKey,
-} from "../use-agent-studio-model-selection-model";
+} from "./model-selection-model";
 
 type ContextUsageCache = {
   externalSessionId: string;
